course project1/requirements.txt
tweepy
pandas
nltk
textblob
confluent_kafka
flask
dash
pymongo
plotly
six